Fixed incorrect enum type used in options argument of addObserver:...
Was using a value from the NSKeyValueChange enumeration, when the argument
type is actually NSKeyValueObservingOptions (a mask, not an enumeration).
Both are typedef-ed to NSUInteger, so the error passed the compiler. The
equivalent value to what was being used would be
NSKeyValueObservingOptionInitial, which does work fine, however, I believe
it is unnecessary, and an options argument of 0 works just as well.